@@ -1,4 +1,4 @@
-from typing import Optional
+from functools import cache
 import networkx as nx
 import pytest
 
@@ -15,31 +15,32 @@ class Solution(ISolution):
         input_data = {d.src: (d.rate, d.dst) for d in inp.get_objects(TTP_TEMPLATE)}
         keys = sorted(input_data)
         self.rates = [input_data[k][0] for k in keys]
-        self.closed_valves = {valve for valve, rate in enumerate(self.rates) if rate > 0}
+        self.closed_valves = [valve for valve, rate in enumerate(self.rates) if rate > 0]
 
         graph = nx.Graph([(key, dst) for key, value in input_data.items() for dst in value[1]])
         distances = nx.floyd_warshall(graph)
         self.costs = [[int(distances[k][v]) + 1 for v in keys] for k in keys]
 
-    def f(self, src_valve, time, time_next: Optional[int]) -> int:
-        if not self.closed_valves:
+    @cache
+    def f(self, valves_state, src_valve, time, time_next: int) -> int:
+        if not valves_state:
             return 0
-        best = self.f(0, time_next, None) if time_next else 0
+        best = self.f(valves_state, 0, time_next, 0) if time_next else 0
 
-        for dst_valve in list(self.closed_valves):
-            if time > (ti := self.costs[src_valve][dst_valve]):
-                self.closed_valves.remove(dst_valve)
-                ti = time - ti
-                best = max(best, ti * self.rates[dst_valve] + self.f(dst_valve, ti, time_next))
-                self.closed_valves.add(dst_valve)
+        for dst_valve in self.closed_valves:
+            if not ((1 << dst_valve) & valves_state):
+                continue
+            if (ti := time - self.costs[src_valve][dst_valve]) > 0:
+                _closed_valves = valves_state - (1 << dst_valve)
+                best = max(best, ti * self.rates[dst_valve] + self.f(_closed_valves, dst_valve, ti, time_next))
 
         return best
 
     def part_a(self):
-        return self.f(0, 30, None)
+        return self.f(sum(1 << valve for valve in self.closed_valves), 0, 30, 0)
 
     def part_b(self):
-        return self.f(0, 26, 26)
+        return self.f(sum(1 << valve for valve in self.closed_valves), 0, 26, 26)
 
 
 @pytest.mark.parametrize("pd", get_puzzles(), ids=str)
